Begin adding different presenters
